requirements.txt
Flask==1.1.1
Flask-RESTX==0.1.1
werkzeug==0.16.1
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.4
pytest==5.3.5
pytest-cov==2.8.1
Flask-Admin==1.5.5